requirements.txt
fastapi==0.115.6
pydantic==2.10.4
requests==2.32.3
python-multipart==0.0.20
python-dotenv==1.0.1
